pipo.fft: changed type of variable new_fft_size from int to long to avoid crash

+++ modules/PiPoFft.h
@@ -185,7 +185,7 @@ public:
   int streamAttributes(bool hasTimeTags, double rate, double offset, unsigned int width, unsigned int height, const char **labels, bool hasVarSize, double domain, unsigned int maxFrames)
   {  
     bool                norm               = norm_attr_.get();
-    int                 new_fft_size       = size_attr_.get();
+    long                new_fft_size       = size_attr_.get();
     enum OutputMode     new_output_mode    = (enum OutputMode)    mode_attr_.get();
     enum WeightingMode  new_weighting_mode = (enum WeightingMode) weighting_attr_.get();
     double              new_samplerate     = (double) height / domain;
